requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
decorator==5.0.7
defusedxml==0.7.1
Django==3.1.7
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-embed-video==1.4.0
django-js-asset==1.2.2
filter==0.0.0.20200724
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
mock==4.0.3
oauth2==1.9.0.post1
oauthlib==3.1.0
Pillow==8.1.2
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
self==2020.12.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
